Create the parts to consume the requestes size in bytes
The type _apart_ is created to contain the random stream of bytes, of the particular size; and a pointer of type _apart_ so a linked list can be created

The function _createParts_ is defined to create the collection of elements of type _apart_ according to the definition stored in the _partCollection_ struct.

A simple method _partCollection.crawl_ is defined to output a text representation of the _partCollection_ struct

A method _partCollection.countPS_ is defined to count the number of elements of a particular size that exist.  It's not being used at the moment.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>
@@ -117,3 +161,34 @@ func fillPart(part []byte) {
 		part[x] = byte(rand.Intn(95) + 32)
 	}
 }
+
+//Create the expected parts, or delete if there are more parts than should from privious runs
+func createParts(ptS *partCollection) {
+	var pap *apart
+	for index,value := range ptS.partSizes {
+		desirednumParts := ptS.partAmmount[index]
+		pap = ptS.partLists[index]
+		if pap == nil && desirednumParts > 0 { //First element
+			var newpart apart
+			newpart.data = make([]byte,value)
+			fillPart(newpart.data)
+			ptS.partLists[index] = &newpart
+			pap = &newpart
+		}
+		for i:=uint64(1); i<desirednumParts; i++ {
+			if pap.next == nil {
+				var newpart apart
+				newpart.data = make([]byte,value)
+				fillPart(newpart.data)
+				pap.next = &newpart
+			}
+			if pap != nil {
+				pap = pap.next
+			}
+		}
+	}
+	if pap !=nil {
+		pap.next = nil
+	}
+}
+