Web service enabled to request the creation of memory parts
Package partmem
===============

The GetActParts function returns information about the actual structure of part that have been already created

GetDefParts returns information about the memory parts computed

The PartCollection data struct is exported so a var based on this type can be used as global variabale in the main program.  The NewpC() creator method is defined to instantiate a PartCollection object.

In DefineParts function, if the number of parts computed for a particular size is zero, the corresponding entry in partAmount array is set to zero, otherwise if a previous run of the method had set a value for that size, it will stick there when a subsequent run computes it to zero.  For the same reason, the loop will run for every size even if the total size has been allocated.

In CreateParts, if the desired number of parts for a particular size is zero, the correspondng first pointer in the partLists array is set to nil, to delete any and all parts for that size.

Main program
============
API points created for:
 /api/mem/add?size=<number of bytes>.- Request the creation of parts for this amount of memory
 /api/mem/getdef.- Get info about the current parts definition
 /api/mem/getact.- Get info about the currect memory reservation

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>