Add concurrency safety meassures.  Replace fatal exits with error messages.
A channel (lock) is used to make sure that a request is not processed while another one is being served.  The lock is used even for getActMem() and getDefMem(), to make sure that the underlying data is not being changed by addMem() while they run.

The channel is used in a non blocking manner, so when a request cannot be served because the lock is taken, it will return a message inmediatelly and finish.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>