Limit for max memory reservation can be specified as env var.
We try to get the value to set the limit for memory requests, for env. var. HIGHMEMLIM.  If the env. var. does not exists or its value is not valid, the value is defaulted to the available free memory in the system, in this case the free memory value is updated on every call to set memory.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>