Added two layer locking mechanism. Adding CONCURRENCY section to REAME.
lock is a number carrying a message, now for a goroutine to be able to procedd needs to have the lock and this has to contain a value of 0.  If the lock is acquired but it contains a number other than 0, the goroutine cannot proceed.  !=0 means that a memory allocation request is being processed.

partmem.createParts() is called as a go routine, to avoid timeouts waiting for it to complete, but this adds the requirement of the "double locking" mechanism, more on this on the README file.
partmem.createParts() blocks waiting for the lock to be available, but a 5 seconds timer is started, if the time passes without getting the lock, the operation is aborted.

The time it takes to process a memory allocation request is counted and printed as a log message

The last memory allocation request timestamp is stored in the PartCollection struct

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>