Created API endpoints for file manipulation.
Created the function GetDefFiles, which counts the number of files and its sizes, according to the definition in the FileCollection struct, not the actual files.
HIGHFILELIM and DATADIR vars created, defining a size limit for the total ammount used by files, and the directory where the files will be created.  These can be created used as environment variables.

Function getfreeDisk defined to obtain the disk space available in the directory where files will be created, this will be used to defined the default HIGHFILELIM variable

Initial version of addFiles created, at the moment it only gets the files definition, but does not create or delete anything.  Does not use locking either.

Initial version of getDefFiles to associate with endpoint /api/disk/getdef.  No locking.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>