The DATADIR env var is assigned a value if empty in the main function, instead of the method addFiles, this way it only has to be done once and the instantiation of the FileCollection object can use that value to define the base dir for file creation
The instantiation of the FileColleciton object initilizes the base directory structure where files will be created.

The function createTree was using an innecesary parameter, dirname, that could be derived from the fileCollection parameter.

HIGHFILELIM is computed everytime the addFiles function is called, because the amount of free disk can and will change with time

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>