Added the logic to create files with random data in response to a /api/disk/add API call that actually requires files to be added
A new function called newFile is responsible for the creation of the files.  The function is called with the name of the file to be created, including the path, and the size in bytes.

To generate the random data, the existing function radstring is used, this function has a limited selection of 26 characters, this reduced spread seems to make it faster than a similar logic in partmem that has a wider spread

The computation of the las file number has been moved up, just after the ordering of the file list

+//Creates a single file of the indicated size
+func newFile(filename string, size uint64) error {
+	var trunch []byte //Trunches of 4k
+	f,err := os.Create(filename)
+	defer f.Close()
+	if err != nil {
+		log.Printf("newFile(): Error creating file: %s",filename)
+		return err
+	}
+	trunch = []byte(randstring(size))
+	_,err = f.Write(trunch)
+	if err != nil {
+		log.Printf("newFile(): Error writing to file: %s",filename)
+		return err
 	}
 	return nil
 }