Creation of directory structure has been moved from the package partdisk to the main package.  Giving a total size of 0 for creating files is legal
Now the directory structure is created at the beginning of the program execution instead of attempting the creation every time the /api/disk/add endpoint is called

The 0 size request for files is equivalent to a "remove all files" request.

A log message with the IP and port where the webserver will listen is sent.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>