Added functionality to request CPU load
A new package "cpuload" has been added to request CPU load through the API endpoint /api/cpu/load?time=<time in senconds>

The time parameter is not honored yet.

The load is achieved by means of factoring a big integer number, at the moment the number used should not take more than 6 to 10 minutes to factor.  That the factoring problem has a finite time to complete is a feature so the load is not added as permanent in case the timeout requested is too high.