Put in place the mechanims to limit the CPU load duration to the time requested by the user
The function LoadUp() will launch the function factor() as a goroutine, and inmediately will wait, using a channel select, for one of two possible outcomes: the time requested by the user runs out; or the number if factorized.  The second outcome is a security meassure to prevent the CPU running high for an indetermine ammount of time.

If the time requested runs out, a message is sent to the factor() function via the quit channel, which is constantly being monitored. If the factor() function reads a value from the quit channel, no matter what the value is, it returns inmediately.

If the factor() function completes the factorization of the number, it will send the factors via a channel to the calling LoadUp() goroutine, which will cause LoadUp() to finish and the cpu lock to be released

A bigger number is used to find its factors, requiring more time (>15min) to complete

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>