Added API endpoint to stop the CPU load.  1 second delays have been added in several places.
The API endpoint /api/cpu/stop is associated with the handler function stopLoad(), this function checks the cpulock.  If the lock is available that means no CPU load request is in progress, so no reason to stop anything.  If the lock is not available then go ahead and try to stop the CPU load.

As security meassure, to stop the load it is required to pass the ID of the current load request.  If the IDs match the load is stopped, otherwise the stop request is rejected.

To be able to cancel the load requests, the definition of locking channels variables have been moved outside the functions.

To finaly cancel the load, the StopLoad() function only needs to inject  a value into the quit channel which is monitored by the factor() function.  When factor reads a value from quit, it signals its calling function LoadUp() and both finish.

On server busy messages and bad data passed on request, a 1 second delay is added in the response to the user, this limits the possibility to  overload the service with bad requests.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>