Support for env var NUMTOFACTOR added.
NUMTOFACTOR env var can be set with an arbitraryly long number, and that is the number that will be used to factor and generate CPU load.  The benefit of this is to pass the around 15 min. factorization time of the default number.

If the NUMTOFACTOR env var does not exist, a default big number will be used.
Default values for NUMTOFACTOR are moved from cpuload package to main package

A new type CpuCollection is defined in the cpuload package to support the NUMTOFACTOR env var and future improvents.  A new instance of this type is initialized from the main function using the NewCc method, which allows the early catching of bad NUMFACTOR definitions like NUMFACTOR="a2345".

If the value assigned to NUMFACTOR is not a valid unsigned integer, the application panics inmediately, instead of waiting for a request for CPU load.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>