New API endpoint to return information about the latest load request.
The API endpoint /api/cpu/getact has been added, no parameters required, when called, if there is a load request in progress, returns the time when the request was made, the duration requested, the time when it will finish (unless the number if factored earlier), and the number used to factorize.

The information is obtained from the CpuCollection type, which now contains an additional field: lapse, with the requested CPU load duration.

Methods to obtain the information from the private fields have been created: GetDuration() and GetReqTime()

The expected finish time is computed and shown in the replay.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>