Modified how random data to inject into files is generated. Added doc section about random data generation.
A new function randbytes() is added to generate the random data to inject into files.  This function uses a simple and quick algorith to generate pseudo random data, it is a lot faster than the rand.Intn() function.  This function returns an array of bytes so the caller function does not have to convert the string to bytes saving memory by reducing the number of data structures.